Issue #961 Move config interface out of api
* CacheEventListenerConfiguration had no place in the API as it is
not a core type
