+++ impl/src/main/java/org/ehcache/config/builders/CacheEventListenerConfigurationBuilder.java
@@ -19,7 +19,7 @@ package org.ehcache.config.builders;
 import org.ehcache.config.Builder;
 import org.ehcache.impl.config.event.DefaultCacheEventListenerConfiguration;
 import org.ehcache.event.CacheEventListener;
-import org.ehcache.event.CacheEventListenerConfiguration;
+import org.ehcache.core.events.CacheEventListenerConfiguration;
 import org.ehcache.event.EventFiring;
 import org.ehcache.event.EventOrdering;
 import org.ehcache.event.EventType;
